+++ src/JIM.Worker/Processors/SyncExportTaskProcessor.cs
@@ -203,8 +203,6 @@ public class SyncExportTaskProcessor
                 ActivityId = _activity.Id,
                 ObjectChangeType = exportItem.ChangeType switch
                 {
-                    PendingExportChangeType.Create => ObjectChangeType.Provisioned,
-                    PendingExportChangeType.Update => ObjectChangeType.Exported,
                     PendingExportChangeType.Delete => ObjectChangeType.Deprovisioned,
                     _ => ObjectChangeType.Exported
                 },